requirements.txt
python-dotenv
SQLAlchemy
geoalchemy2
geopandas
psycopg2-binary
shapely